Add chromium.bbappend: Fixes GCC5.2 induced crash and adds proper library locations
This version of chromium fails to compile under GCC5.2
Meta-browser hacked the build to make it work, but it causes the need to have the flag.
The chromium code has a hard-coded path for the video playing library
Thus it needs to look in the binary path and not the library path.
Added new package which installs the ffmpeg library in the chromium binary path